+++ examples/experimental/components/examples/hello_world_deployment/code_locations/hello_world/hello_world/definitions.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 
+import dagster as dg
 from dagster._components import ComponentRegistry, build_defs_from_toplevel_components_folder
 from dagster._components.impls.pipes_subprocess_script_collection import (
     PipesSubprocessScriptCollection,
 )
-from dagster._core.definitions.definitions_class import Definitions
 
 defs = build_defs_from_toplevel_components_folder(
     path=Path(__file__).parent,
@@ -14,4 +14,4 @@ defs = build_defs_from_toplevel_components_folder(
 )
 
 if __name__ == "__main__":
-    Definitions.validate_loadable(defs)
+    dg.Definitions.validate_loadable(defs)
