+++ python_modules/dagster/dagster/_components/core/deployment.py
@@ -1,0 +1,163 @@
+import importlib.util
+import os
+import sys
+from pathlib import Path
+from typing import Final, Iterable, Type
+
+from typing_extensions import Self
+
+from dagster._components.core.component import (
+    Component,
+    ComponentRegistry,
+    register_components_in_module,
+)
+from dagster._core.errors import DagsterError
+
+
+def is_inside_deployment_project(path: Path) -> bool:
+    try:
+        _resolve_deployment_root_path(path)
+        return True
+    except DagsterError:
+        return False
+
+
+def _resolve_deployment_root_path(path: Path) -> str:
+    current_path = os.path.abspath(path)
+    while not _is_deployment_root(current_path):
+        current_path = os.path.dirname(current_path)
+        if current_path == "/":
+            raise DagsterError("Cannot find deployment root")
+    return current_path
+
+
+def is_inside_code_location_project(path: Path) -> bool:
+    try:
+        _resolve_code_location_root_path(path)
+        return True
+    except DagsterError:
+        return False
+
+
+def _resolve_code_location_root_path(path: Path) -> str:
+    current_path = os.path.abspath(path)
+    while not _is_code_location_root(current_path):
+        current_path = os.path.dirname(current_path)
+        if current_path == "/":
+            raise DagsterError("Cannot find code location root")
+    return current_path
+
+
+def _is_deployment_root(path: str) -> bool:
+    return os.path.exists(os.path.join(path, "code_locations"))
+
+
+def _is_code_location_root(path: str) -> bool:
+    return os.path.basename(os.path.dirname(path)) == "code_locations"
+
+
+# Deployment
+_DEPLOYMENT_CODE_LOCATIONS_DIR: Final = "code_locations"
+
+# Code location
+_CODE_LOCATION_CUSTOM_COMPONENTS_DIR: Final = "lib"
+_CODE_LOCATION_COMPONENT_INSTANCES_DIR: Final = "components"
+
+
+class DeploymentProjectContext:
+    @classmethod
+    def from_path(cls, path: Path) -> Self:
+        return cls(root_path=_resolve_deployment_root_path(path))
+
+    def __init__(self, root_path: str):
+        self._root_path = root_path
+
+    @property
+    def deployment_root(self) -> str:
+        return self._root_path
+
+    @property
+    def code_location_root_path(self) -> str:
+        return os.path.join(self._root_path, _DEPLOYMENT_CODE_LOCATIONS_DIR)
+
+    def has_code_location(self, name: str) -> bool:
+        return os.path.exists(os.path.join(self._root_path, "code_locations", name))
+
+
+class CodeLocationProjectContext:
+    @classmethod
+    def from_path(cls, path: Path, component_registry: "ComponentRegistry") -> Self:
+        root_path = _resolve_code_location_root_path(path)
+        name = os.path.basename(root_path)
+
+        # TODO: Rm when a more robust solution is implemented
+        # Make sure we can import from the cwd
+        if sys.path[0] != "":
+            sys.path.insert(0, "")
+
+        components_lib_module = f"{name}.{_CODE_LOCATION_CUSTOM_COMPONENTS_DIR}"
+        module = importlib.import_module(components_lib_module)
+        register_components_in_module(component_registry, module)
+
+        return cls(
+            deployment_context=DeploymentProjectContext.from_path(path),
+            root_path=root_path,
+            name=os.path.basename(root_path),
+            component_registry=component_registry,
+        )
+
+    def __init__(
+        self,
+        deployment_context: DeploymentProjectContext,
+        root_path: str,
+        name: str,
+        component_registry: "ComponentRegistry",
+    ):
+        self._deployment_context = deployment_context
+        self._root_path = root_path
+        self._name = name
+        self._component_registry = component_registry
+
+    @property
+    def deployment_context(self) -> DeploymentProjectContext:
+        return self._deployment_context
+
+    @property
+    def component_types_root_path(self) -> str:
+        return os.path.join(self._root_path, self._name, _CODE_LOCATION_CUSTOM_COMPONENTS_DIR)
+
+    @property
+    def component_types_root_module(self) -> str:
+        return f"{self._name}.{_CODE_LOCATION_CUSTOM_COMPONENTS_DIR}"
+
+    @property
+    def component_registry(self) -> "ComponentRegistry":
+        return self._component_registry
+
+    def has_component_type(self, name: str) -> bool:
+        return self._component_registry.has(name)
+
+    def get_component_type(self, name: str) -> Type[Component]:
+        if not self.has_component_type(name):
+            raise DagsterError(f"No component type named {name}")
+        return self._component_registry.get(name)
+
+    def get_component_instance_path(self, name: str) -> str:
+        if name not in self.component_instances:
+            raise DagsterError(f"No component instance named {name}")
+        return os.path.join(self.component_instances_root_path, name)
+
+    @property
+    def component_instances_root_path(self) -> str:
+        return os.path.join(self._root_path, self._name, _CODE_LOCATION_COMPONENT_INSTANCES_DIR)
+
+    @property
+    def component_instances(self) -> Iterable[str]:
+        return os.listdir(
+            os.path.join(self._root_path, self._name, _CODE_LOCATION_COMPONENT_INSTANCES_DIR)
+        )
+
+    def has_component_instance(self, name: str) -> bool:
+        return os.path.exists(
+            os.path.join(self._root_path, self._name, _CODE_LOCATION_COMPONENT_INSTANCES_DIR, name)
+        )
