+++ python_modules/dagster/dagster/_generate/templates/CODE_LOCATION_NAME_PLACEHOLDER/CODE_LOCATION_NAME_PLACEHOLDER/definitions.py
@@ -1,6 +1,8 @@
 from pathlib import Path
 
-from dagster._components import build_defs_from_toplevel_components_folder
+from dagster._components.core.component_defs_builder import (
+    build_defs_from_toplevel_components_folder,
+)
 
 defs = build_defs_from_toplevel_components_folder(
     path=Path(__file__).parent,
